Split file collection files table
