+++ src/__tests__/components/file/FileTable.test.tsx
@@ -38,24 +38,6 @@ const pagedPage = {
   status: 200,
 };
 
-const collectionFilesPage = {
-  cursors: { self: "page-1" },
-  data: [
-    {
-      type: "file",
-      id: "file-1",
-      attributes: {
-        name: "File One",
-        status: "complete",
-        suppliedId: "supplied-file-1",
-        created: "2026-06-12T15:30:00Z",
-        uploaded: "2026-06-12T15:31:00Z",
-      },
-    },
-  ],
-  status: 200,
-};
-
 describe("FileTable", () => {
   beforeAll(() => {
     Object.assign(global, {
@@ -152,196 +134,6 @@ describe("FileTable", () => {
     resolveSortedPage?.(page);
     expect(await screen.findByText("alpha.jt")).toBeInTheDocument();
   });
-
-  it("loads collection files from a custom API path while keeping row interactions", async () => {
-    const fetchMock = mockFetch(() => collectionFilesPage);
-    const onFileSelected = jest.fn();
-    jest.spyOn(window, "open").mockReturnValue({} as Window);
-
-    renderTable(onFileSelected, {
-      apiPath: "/api/file-collections/collection-1/files",
-      mode: "filesCollection",
-    });
-
-    expect(await screen.findByText("File One")).toBeInTheDocument();
-    const statusLabel = screen.getByText("complete");
-    expect(statusLabel.closest(".MuiChip-root")).toHaveStyle({
-      textTransform: "uppercase",
-    });
-    expect(fetchMock).toHaveBeenCalledWith(
-      "http://localhost/api/file-collections/collection-1/files?pageSize=25"
-    );
-    expect(screen.getByText("Name")).not.toHaveAttribute("role", "button");
-    expect(
-      screen.queryByLabelText("Supplied ID Filter (exact)")
-    ).not.toBeInTheDocument();
-    expect(
-      screen.queryByRole("button", { name: "New" })
-    ).not.toBeInTheDocument();
-
-    await userEvent.click(screen.getByLabelText("Select File One"));
-
-    expect(screen.getByText("1 selected")).toBeInTheDocument();
-    expect(screen.queryByLabelText("Delete")).not.toBeInTheDocument();
-
-    await userEvent.click(screen.getByText("File One"));
-
-    expect(onFileSelected).toHaveBeenCalledWith({
-      id: "file-1",
-      name: "File One",
-      status: "complete",
-      suppliedId: "supplied-file-1",
-      created: "2026-06-12T15:30:00Z",
-      uploaded: "2026-06-12T15:31:00Z",
-    });
-
-    await userEvent.click(
-      screen.getByRole("button", { name: "Download File One" })
-    );
-
-    await waitFor(() => {
-      expect(fetchMock).toHaveBeenCalledWith("/api/files/file-1/download-url", {
-        method: "POST",
-      });
-    });
-    expect(window.open).toHaveBeenCalledWith(
-      "https://example.test/download/file-1",
-      "_blank",
-      "noopener"
-    );
-  });
-
-  it("disables download for files that are not complete", async () => {
-    const fetchMock = mockFetch(() => ({
-      ...collectionFilesPage,
-      data: [
-        {
-          type: "file",
-          id: "file-1",
-          attributes: {
-            name: "File One",
-            status: "pending",
-            suppliedId: "supplied-file-1",
-            created: "2026-06-12T15:30:00Z",
-            uploaded: "2026-06-12T15:31:00Z",
-          },
-        },
-      ],
-    }));
-
-    renderTable(jest.fn(), {
-      apiPath: "/api/file-collections/collection-1/files",
-      mode: "filesCollection",
-    });
-
-    expect(await screen.findByText("File One")).toBeInTheDocument();
-    expect(screen.getByText("pending")).toBeInTheDocument();
-
-    const download = screen.getByRole("button", { name: "Download File One" });
-    expect(download).toBeDisabled();
-
-    expect(fetchMock).not.toHaveBeenCalledWith(
-      "/api/files/file-1/download-url",
-      {
-        method: "POST",
-      }
-    );
-  });
-
-  it("styles ready file statuses as success", async () => {
-    mockFetch(() => ({
-      ...collectionFilesPage,
-      data: [
-        {
-          type: "file",
-          id: "file-1",
-          attributes: {
-            name: "File One",
-            status: "ready",
-            suppliedId: "supplied-file-1",
-            created: "2026-06-12T15:30:00Z",
-            uploaded: "2026-06-12T15:31:00Z",
-          },
-        },
-      ],
-    }));
-
-    renderTable(jest.fn(), {
-      apiPath: "/api/file-collections/collection-1/files",
-      mode: "filesCollection",
-    });
-
-    const statusLabel = await screen.findByText("ready");
-    expect(statusLabel.closest(".MuiChip-root")).toHaveClass(
-      "MuiChip-colorSuccess"
-    );
-  });
-
-  it("does not treat completed as an available file state", async () => {
-    const fetchMock = mockFetch(() => ({
-      ...collectionFilesPage,
-      data: [
-        {
-          type: "file",
-          id: "file-1",
-          attributes: {
-            name: "File One",
-            status: "completed",
-            suppliedId: "supplied-file-1",
-            created: "2026-06-12T15:30:00Z",
-            uploaded: "2026-06-12T15:31:00Z",
-          },
-        },
-      ],
-    }));
-    jest.spyOn(window, "open").mockReturnValue({} as Window);
-
-    renderTable(jest.fn(), {
-      apiPath: "/api/file-collections/collection-1/files",
-      mode: "filesCollection",
-    });
-
-    expect(await screen.findByText("completed")).toBeInTheDocument();
-
-    const download = screen.getByRole("button", { name: "Download File One" });
-    expect(download).toBeDisabled();
-
-    expect(fetchMock).not.toHaveBeenCalledWith(
-      "/api/files/file-1/download-url",
-      {
-        method: "POST",
-      }
-    );
-  });
-
-  it("renders an empty files table for an empty collection", async () => {
-    mockFetch(() => ({ cursors: { self: "page-1" }, data: [], status: 200 }));
-
-    renderTable(jest.fn(), {
-      apiPath: "/api/file-collections/collection-1/files",
-      mode: "filesCollection",
-    });
-
-    expect(await screen.findByText("Files")).toBeInTheDocument();
-    expect(screen.queryByText("File One")).not.toBeInTheDocument();
-    expect(screen.getByRole("table")).toBeInTheDocument();
-  });
-
-  it("renders load errors with an error indicator", async () => {
-    mockFetch(() => ({
-      message: "Could not load collection files.",
-      status: 500,
-    }));
-
-    renderTable(jest.fn(), {
-      apiPath: "/api/file-collections/collection-1/files",
-      mode: "filesCollection",
-    });
-
-    expect(await screen.findByRole("alert")).toHaveTextContent(
-      "Error loading data."
-    );
-  });
 });
 
 function renderTable(
@@ -362,26 +154,4 @@ function renderTable(
       <FileTable onFileSelected={onFileSelected} {...props} />
     </SWRConfig>
   );
-}
-
-function mockFetch(responseFor: (url: string) => unknown): jest.Mock {
-  const fetchMock = jest.fn((input: RequestInfo | URL) => {
-    const url = input.toString();
-
-    if (url.endsWith("/api/files/file-1/download-url")) {
-      return Promise.resolve({
-        json: () =>
-          Promise.resolve({ url: "https://example.test/download/file-1" }),
-        ok: true,
-      });
-    }
-
-    return Promise.resolve({
-      json: () => Promise.resolve(responseFor(url)),
-      ok: true,
-    });
-  });
-
-  global.fetch = fetchMock as unknown as typeof nodeFetch;
-  return fetchMock;
 }
