+++ src/components/file-collection/FileCollectionFilesTable.tsx
@@ -1,9 +1,6 @@
-import { Add, Download } from "@mui/icons-material";
+import { Download } from "@mui/icons-material";
 import {
   Alert,
-  Box,
-  Button,
-  Checkbox,
   Chip,
   IconButton,
   Paper,
@@ -12,12 +9,11 @@ import {
   TableBody,
   TableCell,
   TableContainer,
+  TableHead,
   TablePagination,
   TableRow,
-  TextField,
   Tooltip,
 } from "@mui/material";
-import debounce from "lodash.debounce";
 import React from "react";
 import useSWR from "swr";
 
@@ -25,47 +21,46 @@ import { isErrorRes } from "../../lib/api";
 import { toLocaleString } from "../../lib/dates";
 import { File, toFilePage } from "../../lib/files";
 import { buildQuery, SwrProps, useCursorPagingState } from "../../lib/paging";
-import { SortState, toggleSort, toSortParam } from "../../lib/sorting";
 import { DataLoadError } from "../shared/DataLoadError";
 import { DefaultPageSize, DefaultRowHeight } from "../shared/Layout";
 import { SkeletonBody } from "../shared/SkeletonBody";
-import { HeadCell, TableHead } from "../shared/TableHead";
 import { TableToolbar } from "../shared/TableToolbar";
-import CreateFileDialog from "./CreateFileDialog";
 
-export const headCells: readonly HeadCell[] = [
-  { id: "name", disablePadding: true, label: "Name", sortable: true },
+interface Props {
+  readonly activeFileId?: string;
+  readonly apiPath: string;
+  readonly onFileSelected: (file: File) => void;
+}
+
+interface UseCollectionFilesProps extends SwrProps {
+  readonly apiPath: string;
+}
+
+const headCells = [
+  { id: "name", label: "Name" },
   { id: "supplied-id", label: "Supplied ID" },
   { id: "status", label: "Status" },
   { id: "id", label: "ID" },
-  { id: "created", label: "Created", sortable: true },
+  { id: "created", label: "Created" },
   { id: "uploaded", label: "Uploaded" },
   { id: "download", label: "Download" },
-];
+] as const;
 
-interface UseFilesProps extends SwrProps {
-  readonly sort?: SortState;
-}
-
-function useFiles({ cursor, pageSize, sort, suppliedId }: UseFilesProps) {
+function useCollectionFiles({
+  apiPath,
+  cursor,
+  pageSize,
+}: UseCollectionFilesProps) {
   return useSWR(
-    buildQuery("/api/files", {
+    buildQuery(apiPath, {
       cursor,
       pageSize,
-      sort: sort != null ? toSortParam(sort) : undefined,
-      suppliedId,
     })
   );
 }
 
 function isFileAvailable(file: File): boolean {
-  const status = normalizeStatus(file.status);
-
-  return status === "complete";
-}
-
-function normalizeStatus(status?: string): string | undefined {
-  return status?.toLowerCase();
+  return file.status?.toLowerCase() === "complete";
 }
 
 function statusLabel(status?: string): string {
@@ -75,7 +70,7 @@ function statusLabel(status?: string): string {
 function statusColor(
   status?: string
 ): "default" | "success" | "warning" | "error" {
-  switch (normalizeStatus(status)) {
+  switch (status?.toLowerCase()) {
     case "complete":
     case "ready":
       return "success";
@@ -89,60 +84,29 @@ function statusColor(
   }
 }
 
-interface Props {
-  readonly activeFileId?: string;
-  readonly onFileSelected: (file: File) => void;
-}
-
-export default function FileTable({
+export default function FileCollectionFilesTable({
   activeFileId,
+  apiPath,
   onFileSelected,
 }: Props): JSX.Element {
   const pageSize = DefaultPageSize;
-  const [sort, setSort] = React.useState<SortState>({
-    field: "created",
-    order: "desc",
-  });
-  const {
-    currentPage,
-    cursor,
-    cursors,
-    handlePageChange,
-    resetPaging,
-    setCursors,
-  } = useCursorPagingState();
-  const [selected, setSelected] = React.useState<Set<string>>(new Set());
-  const [showDialog, setShowDialog] = React.useState(false);
-  const [suppliedId, setSuppliedIdFilter] = React.useState<
-    string | undefined
-  >();
-  const [showToast, setShowToast] = React.useState(false);
+  const { currentPage, cursor, cursors, handlePageChange, setCursors } =
+    useCursorPagingState();
   const [downloadError, setDownloadError] = React.useState<string>();
 
-  const { data, error, mutate } = useFiles({
+  const { data, error } = useCollectionFiles({
+    apiPath,
     cursor,
     pageSize,
-    sort,
-    suppliedId,
   });
   const loadError = error ?? (isErrorRes(data) ? data : undefined);
   const page = data && !isErrorRes(data) ? toFilePage(data) : undefined;
-  const visiblePage = page;
-  const pageLength = visiblePage ? visiblePage.items.length : 0;
-  const paginationCursors = visiblePage?.cursors ?? cursors;
+  const pageLength = page ? page.items.length : 0;
+  const paginationCursors = page?.cursors ?? cursors;
   const emptyRows =
     paginationCursors?.next == null && paginationCursors?.self == null
       ? 0
       : pageSize - pageLength;
-
-  const debouncedSetSuppliedIdFilter = React.useMemo(
-    () =>
-      debounce((value: string) => {
-        resetPaging();
-        setSuppliedIdFilter(value === "" ? undefined : value);
-      }, 300),
-    [resetPaging]
-  );
 
   React.useEffect(() => {
     if (page == null) return;
@@ -150,41 +114,11 @@ export default function FileTable({
     setCursors(page.cursors ?? undefined);
   }, [page, setCursors]);
 
-  function handleSelectAll(e: React.ChangeEvent<HTMLInputElement>) {
-    if (visiblePage == null) return;
-
-    const upd = new Set<string>();
-    if (e.target.checked) visiblePage.items.forEach((n) => upd.add(n.id));
-    setSelected(upd);
-  }
-
-  function handleCheck(id: string) {
-    const upd = new Set(selected);
-    if (selected.has(id)) upd.delete(id);
-    else upd.add(id);
-
-    setSelected(upd);
-  }
-
   function handleChangePage(
     _: React.MouseEvent<HTMLButtonElement> | null,
     num: number
   ) {
     handlePageChange(num);
-  }
-
-  function handleSortChange(field: string) {
-    setSort((current) => toggleSort(current, field));
-    resetPaging();
-  }
-
-  async function handleDelete() {
-    setSelected(new Set());
-    await fetch("/api/files", {
-      body: JSON.stringify({ ids: [...selected] }),
-      method: "DELETE",
-    });
-    mutate();
   }
 
   async function handleDownload(id: string) {
@@ -213,47 +147,30 @@ export default function FileTable({
 
   let tableRows: React.ReactNode;
   if (loadError) {
-    tableRows = <DataLoadError colSpan={headCells.length + 1} />;
-  } else if (!visiblePage) {
+    tableRows = <DataLoadError colSpan={headCells.length} />;
+  } else if (!page) {
     tableRows = (
       <SkeletonBody
-        includeCheckbox={true}
-        numCellsPerRow={8}
+        includeCheckbox={false}
+        numCellsPerRow={headCells.length}
         numRows={pageSize - pageLength}
         rowHeight={DefaultRowHeight}
       />
     );
   } else {
-    tableRows = visiblePage.items.map((row) => {
-      const isSel = selected.has(row.id);
+    tableRows = page.items.map((row) => {
       const isActive = activeFileId === row.id;
       const isAvailable = isFileAvailable(row);
 
       return (
         <TableRow
           hover
-          role="checkbox"
           tabIndex={-1}
           key={row.id}
-          selected={isSel || isActive}
+          selected={isActive}
           onClick={() => onFileSelected(row)}
         >
-          <TableCell
-            padding="checkbox"
-            onClick={(e) => {
-              e.stopPropagation();
-              handleCheck(row.id);
-            }}
-          >
-            <Checkbox
-              color="primary"
-              checked={isSel}
-              inputProps={{
-                "aria-label": `Select ${row.name ?? row.id}`,
-              }}
-            />
-          </TableCell>
-          <TableCell component="th" scope="row" padding="none">
+          <TableCell component="th" scope="row">
             {row.name}
           </TableCell>
           <TableCell>{row.suppliedId}</TableCell>
@@ -302,55 +219,21 @@ export default function FileTable({
   return (
     <>
       <Paper sx={{ m: 2 }}>
-        <TableToolbar
-          numSelected={selected.size}
-          onDelete={handleDelete}
-          title="Files"
-        />
-        <Box
-          sx={{
-            px: { sm: 2 },
-            display: "flex",
-            justifyContent: "space-between",
-            alignItems: "center",
-          }}
-        >
-          <TextField
-            variant="standard"
-            size="small"
-            margin="normal"
-            id="suppliedIdFilter"
-            label="Supplied ID Filter (exact)"
-            type="text"
-            onChange={(e) => {
-              debouncedSetSuppliedIdFilter(e.target.value?.trim() ?? "");
-            }}
-            sx={{ mt: 0, width: "20rem" }}
-          />
-          <Button
-            key="upload"
-            variant="contained"
-            startIcon={<Add />}
-            onClick={() => setShowDialog(true)}
-          >
-            New
-          </Button>
-        </Box>
+        <TableToolbar numSelected={0} title="Files" />
         <TableContainer>
           <Table>
-            <TableHead
-              headCells={headCells}
-              numSelected={selected.size}
-              onSelectAllClick={handleSelectAll}
-              onSortChange={handleSortChange}
-              rowCount={pageLength}
-              sort={sort}
-            />
+            <TableHead>
+              <TableRow>
+                {headCells.map((headCell) => (
+                  <TableCell key={headCell.id}>{headCell.label}</TableCell>
+                ))}
+              </TableRow>
+            </TableHead>
             <TableBody>
               {tableRows}
               {emptyRows > 0 && (
                 <TableRow style={{ height: DefaultRowHeight * emptyRows }}>
-                  <TableCell colSpan={headCells.length + 1} />
+                  <TableCell colSpan={headCells.length} />
                 </TableRow>
               )}
             </TableBody>
@@ -370,24 +253,6 @@ export default function FileTable({
           }}
         />
       </Paper>
-      <CreateFileDialog
-        open={showDialog}
-        onClose={() => setShowDialog(false)}
-        onFileCreated={() => {
-          setShowDialog(false);
-          setShowToast(true);
-          mutate();
-        }}
-      />
-      <Snackbar
-        open={showToast}
-        autoHideDuration={6000}
-        onClose={() => setShowToast(false)}
-      >
-        <Alert onClose={() => setShowToast(false)} severity="success">
-          File created!
-        </Alert>
-      </Snackbar>
       <Snackbar
         open={downloadError != null}
         autoHideDuration={6000}
